pyproject.toml
[tool.poetry]
name = "7_automated_emails"
version = "0.1.0"
description = ""
authors = ["martinstajnko <stajnko.martin93@gmail.com>"]

[tool.poetry.dependencies]
python = "3.9.5"
requests = "^2.26.0"
beautifulsoup4 = "^4.10.0"
python-dotenv = "^0.19.1"
yagmail = "^0.14.260"
pandas = "^1.3.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
